refactor: move base axios to libs folder
